IntelHex started handling record types of: 0, 1, 2

+++ Arselon.Cdt.Hex/BinaryMap.cs
@@ -8,7 +8,7 @@ namespace Arselon.Cdt.Hex
 {
     public class BinaryMap : IBinaryMap
     {
-        public void AddData(uint address, byte[] data)
+        public void AddData(long address, byte[] data)
         {
         }
     }
